feat: Refactor Fourthwall webhook handling and introduce OrderSyncService for better order management
